tests/wpt/tests/encoding-detection/uk-KOI8-U-late.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>uk KOI8-U late</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "KOI8-U", 'Expected KOI8-U');
    }, "Check detection result");
    w.close();
    done();
};
var w = window.open("support/uk-KOI8-U-late.sub.html");
</script>
