third_party/blink/web_tests/external/wpt/encoding-detection/uk-KOI8-U-late.tentative-expected.txt

This is a testharness.js-based test.
FAIL Check detection result assert_equals: Expected KOI8-U expected "KOI8-U" but got "KOI8-R"
Harness: the test ran to completion.
